worker-service/requirements.txt
psycopg2-binary==2.9.7
redis==4.5.4
python-dotenv==1.0.0